refactor: isolate delegated expansion execution
